refactor: ContentList의 중요 로직 분리 및 분기
